[svn] Fixed problem with md5  (long -> int)

+++ CDL/Util/md5.h
@@ -39,7 +39,7 @@ typedef unsigned char *POINTER;
 typedef unsigned short int UINT2;
 
 /* UINT4 defines a four byte word */
-typedef unsigned long int UINT4;
+typedef unsigned int UINT4;
 
 
 typedef struct {
